Move useSyncExternalStore tests to separate file
